backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[bcrypt]
python-jose
pydantic
email-validator
